issue fixed
git-svn-id: svn://127.0.0.1/flaskdump/trunk@2524 386f392f-2c41-1343-9bf7-4b6fedb51b34

+++ mobileapp/FlaskUs/www/js/events/manage_event_content.js
@@ -211,12 +211,23 @@
             $scope.editContent=true;
         }
         $scope.deleteEventDetail=function(eventDetailId,index){
-            EventsService.deleteEventDetailById(eventDetailId).then(function(response){
-                if(response.data){
-                    $scope.flaskUsDetails.splice(index,1);
-                }
-
+            var confirmPopup = $ionicPopup.confirm({
+                title: 'Delete Detail?',
+                template: 'Select OK to Confirm!'
             });
+            confirmPopup.then(function(res) {
+                if(res) {
+                    EventsService.deleteEventDetailById(eventDetailId).then(function(response){
+                        if(response.data){
+                            $scope.flaskUsDetails.splice(index,1);
+                        }
+
+                    });
+
+                } else {
+                }
+            });
+
 
         }
         $scope.initialize=function() {
